refactor(#release): standardize camel case file naming
