--- packages/release/src/cli/actions/prepare-release/index.ts
@@ -1,3 +1,0 @@
-import prepareRelease from './prepare-release';
-
-export default prepareRelease;
