+++ packages/release/src/cli/actions/prepareRelease/index.ts
@@ -1,0 +1,3 @@
+import prepareRelease from './prepareRelease';
+
+export default prepareRelease;
