+++ packages/release/src/config/releaseConfig/__tests__/fixtures/invalid/syntaxError/release.config.js
