+++ packages/release/src/config/releaseConfig/__tests__/fixtures/invalid/validationError/release.config.js
