+++ packages/release/src/config/releaseConfig/__tests__/fixtures/valid/release.config.js
